OfficeBeacon/requirements.txt
django==1.11
django-rest_framework
django-rest_auth
